+++ G-Shift/G-Shift/Enemy1a.cs
@@ -35,7 +35,7 @@ namespace G_Shift
         //public int health;
         public int health { get; set; }
 
-        public int depth { get; set; }
+        public float depth { get; set; }
 
         public Rectangle rect
         {
