+++ G-Shift/G-Shift/Enemy2b.cs
@@ -36,7 +36,7 @@ namespace G_Shift
         //float size;
         public int ttl { get; set; }
         public int health { get; set; }
-        public int depth { get; set; }
+        public float depth { get; set; }
 
         public bool jumpFlag { get; set; }
         public float gravity;
@@ -102,7 +102,7 @@ namespace G_Shift
         public Texture2D spriteSheetmoveLeft;
         private AnimatedSprite moveLeftAnimation;
         private AnimatedSprite moveRightAnimation;
-
+        //public float depth { get; set; }
         public enum Stance
         {
             Wait,
